+++ tests/Neo.Plugins.L2Gateway.UnitTests/Usings.cs
@@ -1,3 +1,4 @@
 global using Microsoft.VisualStudio.TestTools.UnitTesting;
 global using Neo.L2;
+global using Neo.L2.Telemetry;
 global using Neo.Plugins.L2Gateway;